Move CalcException from header to its only consumer
Removes the now-empty header as well
