+++ src/CalcManager/CEngine/History.cpp
@@ -6,12 +6,26 @@
 #include "Command.h"
 #include "CalculatorVector.h"
 #include "ExpressionCommand.h"
-#include "CalcException.h"
 
 constexpr int ASCII_0 = 48;
 
 using namespace std;
 using namespace CalcEngine;
+
+class CalcException : public std::exception
+{
+public:
+    CalcException(HRESULT hr)
+    {
+        m_hr = hr;
+    }
+    HRESULT GetException()
+    {
+        return m_hr;
+    }
+private:
+    HRESULT m_hr;
+};
 
 namespace {
     void IFT(HRESULT hr)
